Eliminate "ghost-square" hack for hidden squares
Previously, I was replacing hidden squares (e.g. a square to the right
of one with a horizontal domino) with ghost squares, because they were
otherwise hiding half of the domino. This is now fixed with correct use
of z-index.

A consequence of this is that the domino need not cover the entire
squares. A more natural, slightly rounded off shape, might be better (it
will be easier to distinguish between close-together dominoes, and also,
might allow for clearer insight of the fact that each domino covers two
different colors). Before, this was not possible because one of the
squares "underneath" the domino did not actually exist.