Fix fast timer bug, add topics test instructions
The interval for the timer was not being cleared when changing pages,
and initiating another interval would make it count down twice as
fast. This might actually be a non-issue in the final app, if we only
allow moving forward once the time is up anyway, but good to fix for
now.

Also, we now have the instructions and basic layout for topics test.