Change hint alert mechanism
The alert action was troublesome with the react-dnd library (alerting
during a drag action seemed to put the drag into a limbo and disrupt the
whole thing). Now, hints are made known via an overlay notification made
in pure react instead.

Also, slightly modified the "timesUp" update for tests, since using
setState multiple times in a single function must follow some rules. I
don't think this was an issue for me specifically, but I thought it
better to be safe than sorry.