Make one more instruction screen in checkerboard
Now the warning that "You might find using the interface is clunky and
slow" is on a screen of its own in the tour of the checkerboard
interface.

Change "be very thorough in recording your thoughts" to "be very liberal
in recording your thoughts" (in two occasions), so as to not give the
participant the impression that they should give many details.

Make instructions for how to enter the answer to phase 2 of experiment
more explicit (they should be entered in the chat box).

Change "anything else which" to "anything else that"