Make small changes to tests pages
The instruction page for Nonsense Syllogisms now explicitly says each of
the two parts contains 15 questions.

The instruction for what each option means is not repeated in the actual
test pages of the Nonsense Syllogisms test, after the instruction page.

The styling for form questions now has a smaller bottom padding, so more
test quastions can fit in the screen.