Update Questionnaire questions
Changes to all pages of questionnaire:

Page 1: Addition of very general initial question, and collapse of second and
third questions.

Page 2: Addition of "by the time I was told it is impossible, I was
already convinced it was impossible" question, and of "This question
refers to..." markers.

Page 3: Addition of "This question refers to..." markers.